refactor(conventions): rename interface-utils to interface-ast-predicates
Removes forbidden generic name "utils" per project conventions.
