+++ packages/riviere-extract-conventions/src/eslint/event-handler-requires-subscribed-events.cjs
@@ -3,7 +3,7 @@ const {
   findInstanceProperty,
   hasLiteralArrayValue,
   getValueTypeDescription,
-} = require('./interface-utils.cjs')
+} = require('./interface-ast-predicates.cjs')
 
 module.exports = {
   meta: {
